# src/dist/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Example API"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    }
  },
  "paths": {
    "/api/user/signup": {
      "post": {
        "summary": "Create a user",
        "requestBody": {
          "required": "true",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "mail": {
                    "type": "string",
                    "format": "email"
                  },
                  "surname": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["name", "mail", "surname", "password"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creation successful"
          },
          "400": {
            "description": "Unable to signup"
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "summary": "Login a user",
        "requestBody": {
          "required": "true",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mail": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["mail", "password"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Unable to login"
          }
        }
      }
    },
    "/api/user/me": {
      "get": {
        "summary": "Return my user",
        "security": {
          "cookieAuth": []
        },
        "responses": {
          "200": {
            "description": "Return my user info"
          },
          "401": {
            "description": "User not logged in"
          }
        }
      }
    }
  },
  "definitions": {
    "userSignupRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "mail": {
          "type": "string",
          "format": "email"
        },
        "surname": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["name", "mail", "surname", "password"],
      "additionalProperties": false
    },
    "userLoginRequest": {
      "type": "object",
      "properties": {
        "mail": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["mail", "password"],
      "additionalProperties": false
    }
  }
}